fix(sglang): harden mocker gRPC server
