+++ CODEOWNERS
@@ -511,6 +511,7 @@
 /lib/llm/src/preprocessor/                                                   @ai-dynamo/dynamo-frontend-codeowners
 /tests/gpu_memory_service/                                                   @ai-dynamo/dynamo-gms-codeowners
 /lib/llm/src/block_manager/                                                  @ai-dynamo/dynamo-kv-memory-codeowners
+/lib/mocker/servers/sglang/                                                  @ai-dynamo/dynamo-backend-sglang-codeowners
 /lib/runtime/src/metrics.rs                                                  @ai-dynamo/dynamo-observability-codeowners
 /tests/vllm_self_benchmark/                                                  @ai-dynamo/dynamo-performance-codeowners
 /deploy/pre-deployment/nixl/                                                 @ai-dynamo/dynamo-kv-memory-codeowners
@@ -605,6 +606,7 @@
 /lib/runtime/src/nvtx.rs                                                     @ai-dynamo/dynamo-runtime-codeowners @ai-dynamo/dynamo-performance-codeowners
 /tests/gpu_memory_service/                                                   @ai-dynamo/dynamo-gms-codeowners
 /lib/llm/src/request_trace/                                                  @ai-dynamo/dynamo-agents-codeowners @ai-dynamo/dynamo-observability-codeowners
+/lib/mocker/servers/sglang/                                                  @ai-dynamo/dynamo-performance-codeowners @ai-dynamo/dynamo-backend-sglang-codeowners
 /lib/mocker/src/kv_manager/                                                  @ai-dynamo/dynamo-performance-codeowners @ai-dynamo/dynamo-kv-memory-codeowners
 /lib/runtime/src/logging.rs                                                  @ai-dynamo/dynamo-runtime-codeowners @ai-dynamo/dynamo-observability-codeowners
 /benchmarks/multimodal/image/                                                @ai-dynamo/dynamo-multimodal-codeowners @ai-dynamo/dynamo-performance-codeowners
