+++ lib/mocker/servers/sglang/tests/sidecar.rs
@@ -211,7 +211,6 @@ async fn prefill_handoff_round_trips_through_a_decode_server() {
     assert_eq!(handoff["bootstrap_port"], 8_998);
     assert!(handoff["bootstrap_room"].is_number());
 
-    let expected_handoff = handoff.clone();
     let mut decode_request = request(3);
     decode_request.prefill_result = Some(PrefillResult {
         disaggregated_params: handoff,
@@ -222,22 +221,6 @@ async fn prefill_handoff_round_trips_through_a_decode_server() {
     assert_eq!(
         decode_outputs.last().unwrap().finish_reason,
         Some(FinishReason::Length)
-    );
-    let observed = decode_server
-        .service
-        .last_disaggregated_params()
-        .expect("decode server should observe rendezvous metadata");
-    assert_eq!(
-        observed.bootstrap_host,
-        expected_handoff["bootstrap_host"].as_str().unwrap()
-    );
-    assert_eq!(
-        i64::from(observed.bootstrap_port),
-        expected_handoff["bootstrap_port"].as_i64().unwrap()
-    );
-    assert_eq!(
-        observed.bootstrap_room,
-        expected_handoff["bootstrap_room"].as_i64().unwrap()
     );
 }
