+++ lib/sidecar/sglang/src/protocol.rs
@@ -434,9 +434,24 @@ fn prompt_logprobs_from_meta(meta: &HashMap<String, String>) -> Result<Option<Va
         _ => Vec::new(),
     };
 
-    let mut payload = Vec::with_capacity(input_logprobs.len() + 1);
-    payload.push(Value::Null);
+    // Current SGLang encodes the first requested prompt position as
+    // `[null, token_id, null]`; older releases omitted it entirely.
+    let has_native_sentinel = input_logprobs.first().is_some_and(|entry| {
+        entry.is_null()
+            || entry
+                .as_array()
+                .and_then(|parts| parts.first())
+                .is_some_and(Value::is_null)
+    });
+    let mut payload = Vec::with_capacity(input_logprobs.len() + usize::from(!has_native_sentinel));
+    if !has_native_sentinel {
+        payload.push(Value::Null);
+    }
     for (index, selected) in input_logprobs.iter().enumerate() {
+        if index == 0 && has_native_sentinel {
+            payload.push(Value::Null);
+            continue;
+        }
         let (token_id, entry) = prompt_logprob_entry(selected, "input_token_logprobs")?;
         let mut position = Map::new();
         position.insert(token_id, entry);
@@ -740,18 +755,18 @@ mod tests {
         let meta = HashMap::from([
             (
                 "input_token_logprobs".to_string(),
-                json!([[-0.1, 10, "a"], [-0.2, 11, "b"]]).to_string(),
+                json!([[null, 10, null], [-0.2, 11, "b"]]).to_string(),
             ),
             (
                 "input_top_logprobs".to_string(),
-                json!([[[-0.3, 12, "c"]], []]).to_string(),
+                json!([null, [[-0.3, 12, "c"]]]).to_string(),
             ),
             ("routed_experts".to_string(), json!([1, 2]).to_string()),
         ]);
         let data = engine_data_from_meta(&meta, true).unwrap().unwrap();
         let prompt = data["prompt_logprobs"].as_array().unwrap();
         assert!(prompt[0].is_null());
-        assert_eq!(prompt[1]["10"]["logprob"], json!(-0.1));
+        assert_eq!(prompt[1]["11"]["logprob"], json!(-0.2));
         assert_eq!(prompt[1]["12"]["decoded_token"], json!("c"));
         assert_eq!(data["routed_experts"], json!([1, 2]));
     }
